Improvements after code review
* union ids in a set instead of browse for performance in case of many
  moves
* improve summary description in glue module
* remove redundant code in test
* split test common case in "common.py": we shouldn't import a test file
  from another module
